+++ frontend/src/App.svelte
@@ -1,9 +1,7 @@
 <script>
   import FetchBudget from './Components/FetchBudget.svelte'
-  import Button from './Components/Button.svelte'
   import Nav from './Components/Nav.svelte'
-  import Acc from './Components/acc.svelte';
-  // import { budget } from './Components/FetchBudget.svelte'
+  import Accordion from './Components/Accordion.svelte';
   let budget = { expenses: [], income: [] };
 </script>
 
@@ -15,7 +13,7 @@
   <Nav />
   <div class="wrapper">
     <div class="content-first">
-      <Acc { budget }/>
+      <Accordion { budget }/>
     </div>
     <div class="content-second">
