--- frontend/src/Components/acc.svelte
@@ -1,170 +1,0 @@
-<script>
-  import Cross from "../assets/circle-x.svg";
-  import Add from "../assets/circle-plus.svg";
-  import PlusIcon from "../assets/plus.svg";
-  import MinusIcon from "../assets/minus.svg";
-  import { GetLatestIdOfItem } from "./FetchLatestId.svelte";
-  export let budget;
-  let indexOf;
-  let openAccordionIndex = null;
-
-
-  $: errorResponse = null;
-
-  const AddItem = (category) => () => {
-    let itemToAdd = {
-      id: GetLatestIdOfItem(budget),
-      name: "New Item",
-      totalAmount: 0,
-    };
-    indexOf = budget.expenses.findIndex((cat) => cat.name == category);
-    budget.expenses[indexOf].items = [
-      ...budget.expenses[indexOf].items,
-      itemToAdd,
-    ];
-    console.log(itemToAdd.id);
-    console.log(budget);
-  };
-
-  const DeleteItem = (categoryIndex, itemId) => {
-    var itemIndex = budget.expenses[categoryIndex].items.findIndex(
-      (item) => item.id == itemId
-    );
-    if (itemIndex !== -1) {
-      budget.expenses[categoryIndex].items.splice(itemIndex, 1);
-      budget = { ...budget };
-    }
-  };
-
-    $: budget.expenses.forEach((expense) => {
-    let totalAmount = 0;
-    expense.items.forEach((item) => {
-      totalAmount += item.amount;
-    });
-    expense.totalAmount = totalAmount;
-  });
-
-  $: budget.expenses.forEach((expense) => {
-    let totalAmount = 0;
-    expense.items.forEach((item) => {
-      totalAmount += parseFloat(item.amount || 0);
-    });
-    expense.totalAmount = totalAmount;
-  });
-
-
-  $: totalAmountExpense = budget.expenses.reduce(
-    (total, expense) => total + expense.totalAmount,
-    0
-  );
-
-  function toggleAccordion(index) {
-    if (openAccordionIndex === index) {
-      openAccordionIndex = null;
-    } else {
-      openAccordionIndex = index;
-    }
-  }
-</script>
-
-<style>
-  .accordion-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    cursor: pointer;
-    background-color: #091f20;
-    border: 2px solid #091f20;
-    border-radius: 4px;
-    color: #dff4f6;
-    padding: 1rem 1rem;
-  }
-
-  .accordion-header-open {
-    border-bottom-left-radius: 0;
-    border-bottom-right-radius: 0;
-  }
-
-  .accordion-content {
-    display: none;
-  }
-
-  .accordion-content.open {
-    display: block;
-    text-align: center;
-  }
-</style>
-
-
-<!-- Markup för Accordion-->
-
-{#each budget.expenses as expense, index}
-  <div
-    role="button"
-    tabindex="0" 
-    class="accordion-header subdisplay mt-3 {openAccordionIndex === index ? 'accordion-header-open' : ''}"
-    on:click={() => toggleAccordion(index)}
-    on:keydown={(e) => e.key === 'Enter' && toggleAccordion(index)}
-    aria-expanded={openAccordionIndex === index ? 'true' : 'false'}
-  >
-    <span>
-      {expense.name}
-      {#if expense.totalAmount && expense.totalAmount !== 0}
-        - {expense.totalAmount}
-      {/if}
-    </span>
-
-    {#if openAccordionIndex === index}
-      <img src={MinusIcon} alt="Collapse" class="accordion-icon" />
-    {:else}
-      <img src={PlusIcon} alt="Expand" class="accordion-icon" />
-    {/if}
-  </div>
-
-  {#if openAccordionIndex === index}
-    <div class="accordion-bg-color">
-      <div class="accordion-content-first">
-        <p class="accordion-paragraph">
-          Lorem ipsum dolor sit amet consectetur adipisicing elit. Iste
-          veritatis inventore repellat recusandae minus itaque, a porro
-          similique soluta facilis non omnis laudantium, impedit eum
-          fugit? Vitae veniam sint quidem!
-        </p>
-      </div>
-      {#each expense.items as item}
-        <div class="accordion-wrapper mt-2">
-          <div class="accordion-content-first">
-            <button
-              class="icon-button"
-              on:click={() => DeleteItem(index, item.id)}
-            >
-              <img
-                src={Cross}
-                class="item-icons"
-                alt="delete itemfield"
-              />
-            </button>
-            <input
-              class="accordion-item-name"
-              type="text"
-              bind:value={item.name}
-            />
-            <input
-              class="accordion-item-amount mx-2"
-              type="number"
-              min="0"
-              bind:value={item.amount}
-            />
-          </div>
-        </div>
-    {/each}
-      <div class="accordion-full-bleed">
-        <button class="icon-button mt-4" on:click={AddItem(expense.name)}>
-          <img src={Add} class="item-icons" alt="Add item" />
-          <p class="small-p">Add new field</p>
-        </button>
-      </div>
-    </div>
-  {/if}
-{/each}
-<!-- Markup -->
